+++ config/services.php
@@ -19,6 +19,7 @@ use Sofascore\PurgatoryBundle\Cache\TargetResolver\ForResponseGroupsResolver;
 use Sofascore\PurgatoryBundle\Command\DebugCommand;
 use Sofascore\PurgatoryBundle\Doctrine\DBAL\Middleware;
 use Sofascore\PurgatoryBundle\Listener\EntityChangeListener;
+use Sofascore\PurgatoryBundle\Listener\EntityChangePurgeSwitcher;
 use Sofascore\PurgatoryBundle\Purger\AsyncPurger;
 use Sofascore\PurgatoryBundle\Purger\InMemoryPurger;
 use Sofascore\PurgatoryBundle\Purger\Messenger\PurgeMessageHandler;
@@ -166,11 +167,14 @@ return static function (ContainerConfigurator $container) {
             ->tag('purgatory.route_provider')
             ->arg(3, service('sofascore.purgatory.property_accessor'))
 
+        ->set('sofascore.purgatory.entity_change_purge_switcher', EntityChangePurgeSwitcher::class)
+
         ->set('sofascore.purgatory.entity_change_listener', EntityChangeListener::class)
             ->args([
                 tagged_iterator('purgatory.route_provider'),
                 service('router'),
                 service('sofascore.purgatory.purger'),
+                service('sofascore.purgatory.entity_change_purge_switcher'),
             ])
 
         ->set('sofascore.purgatory.purger.void', VoidPurger::class)