+++ tests/PHPUnit/Fixtures/WithInheritedClassAttributeDummy.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Sofascore\PurgatoryBundle\Tests\PHPUnit\Fixtures;
+
+final class WithInheritedClassAttributeDummy extends WithClassAttributeDummy
+{
+    public function testInherited(): void
+    {
+    }
+}